Support custom send and receive function
